Use RelSystemPath & Uri types
RetargetablePathBuf was removed in favor of RelSystemPath. RelSystemPath
has two components: a "base" and a "target". The "base" is the relative
portion of the path and can be easily swapped out for a different base.
This allows RelSystemPath to be used for representing a source file on
the system, and then switching the base to represent a target file on
the system. So we can go from a base of "my_site/src" to
"my_site/output" and the target file remains the same.

Additional helper methods are present to rename the target file, change
file extensions, and changing the parent of the target file.

The Uri type was also added which replaced CanonicalPath. Uri
(currently) never starts with a slash (/) and is always relative to the
root of the site. Uri is also used for the internal search key for pages
and is exposed to scripting. If a user wants to query a specific page,
they can use the Uri of the page from the perspective of the browser. So
if we have a source file of "my_site/src/index.md", a script can just
use "index.md" to target this file.

Use Uri for search keys

Add tests for RelSystemPath and Uri

Refactor

Move page constructor logic into free functions

Allow creation of Page from a Reader

Add tests for Page and RelSystemPath

Use macros in test

Refactor

Add test for PageStore

Add HTML rewriter and asset discovery

Original implementation only read HTML and performed the necessary
operations to generate target files. The problem with this was pages
could be moved into subdirectories in order to generate "index" links.
This means that all relative asset paths broke because the output page
no longer matched the source tree.

Assets in HTML will no be rewritten to use absolute paths when
applicable.

Add tests for each handled HTML tag

Add test assets