Properly handle pages named index
When a page is marked as `use_index` and the page is also named "index",
then there was an issue where a subdirectory named "index" was created.
This is expected behavior for the `use_index` flag, but doesn't make
sense because a page named "index" is already the index. This is now
taken into account and when encountered, the html tag rewriter will
properly reference the correct location.