Fix panic when handling many requests
When multiple requests come in from the same user on the devserver, the
original request would get dropped and a new one would be sent. Dropping
the initial request would close the communication channel between the
devserver and the engine. This caused a panic in the engine where
`.expect` was used. The `.expect` call was changed to log a warning
instead.