Change pipeline shell tokens
Added a `$SCRATCH` and `$NEW_SCRATCH` tokens to shell pipeline
processing. `$NEW_SCRATCH` will create a new temp file to be used for
processing. `$SCRATCH` refers to the temp file that was created in the
previous `$NEW_SCRATCH` command.

Also removed the requirement of appending a `COPY` operation to the end
of the pipeline. When using shell commands, the copy will happen
automatically.