Add a RenderBehavior option to server
Some tooling requires the rendered HTML to exist on disk for processing.
An option named `RenderBehavior` was added that will write the page to
disk. This write will occur prior to processing pipelines, so the HTML
page will be available for pipeline processing. The default behavior is
set to `RenderBehavior::Write`, which will write the page on each
request. Use `RenderBehavior::Memory` to only serve the pages from
memory.