Pipelines now have a base directory
Previously it was assumed that everything was in the "content"
directory, for pipeline processing but this assumption fails
when dealing with mounted directories.

This change fixes the problem by allowing the "base" directory for
pipeline processing to be set when creating the pipeline. So instead of
assuming "content", the user can set where to search for files. This
allows having a "static" directory or "wwwroot" directory separate from
"content".

Note that this change only applies when using `$SOURCE` in a shell
command, or when using the `Operation::Copy` command.

If using a shell command without `$SOURCE`, then the base will
not be used and it is expected that the user will define exactly
where files reside. This happens if there is a single asset being
processed by the pipeline. Examples: global CSS, static images.