Completely rewrite path system
The new path system has "Checked" and "Unchecked" types.

The "Checked" types can only be created if the target exists. They are
also generic over any "marker" types, which are used to leverage
compiler type checking. These checks will make sure a path is (for
example) only targeting a `.html` file in the output directory, or
only a `.md` file in the source directory.

"Unchecked" types can be created at-will and are mainly used for path
manipulation. They have "Relative" and "Absolute" versions, which
indicate if the underlying path is relative or absolute. They have
`join` methods that only allow joining when appropriate. "Unchecked"
types can be converted into "Checked" types using any marker and the
conversion will only succeed if the target exists and passes the
conditions in the `CheckedPath` trait.