Remove assumed watch targets
Only the site-rules script, mounts, and watches indicated in the
site-rules script are now watched. Previously, it was assumed that
`content/` and `web/` and `syntax-themes/` were always watched. Not
all sites will use `syntax-themes/`, so instead all assumed watches
were removed.

`web/` and `content/` are now watched manually in the site-rules script
instead of being assumed, and the default script that ships with Pylon
now includes these directores in the site-rules script.