Workaround for key lookup in Tera templates
Having both a `/` and `.` as a map key causes rendering issues with
Tera. As a workaround, the `.md` part of the path was removed from the
keys. Tera templates can use the `replace` filter to remove `.md` from
target documents in shortcodes and templates.

This was added because passing in variables into macros or shortcodes to
use as a key in a library lookup were failing. Hard-coded paths did not
exhibit this issue.