Update internal logic for doc search in context
Previously, the path was used as a search key, but the extension was
removed. This could be misleading, so the extension was added back to
the path and a new field was created: search_key. The search_key field
is the relative path of the document, minus the extension. The extension
is removed due to a bug in Tera, which breaks when an object is being
access by key and the key contains both a `/` and `.`.