Autocopy non-pipelined assets.
All assets that do not have a defined pipeline will have a default
"copy" pipeline created and executed. This allows for assets to be
automatically copied without the need to specify them in the
site-rules file.

The behavior only supports copying colocated assets that are relative to
the directory in which the document that triggered the asset lookup
exists: `content/index.md` requests `test.png` then a copy will be
attempted from `content/test.png`.