Autocopy colocated assets
Pipeline procesing was split into multiple steps: first a pipeline run,
and then an asset colocation run. The asset colocation was modified to
allow assets that are not available during the first run to be properly
discovered at the end of processing. This enables documents to link to
colocated assets in other documents, provided they use an absolute URI.